#24 [REFACTOR] prompt 파일 수정
