+++ src/prompts/controllers/prompt.controller.ts
@@ -1,9 +1,9 @@
 import { Request, Response } from "express";
-import { SearchPromptDto } from "./dtos/search-prompt.dto";
-import * as promptService from "./prompt.service";
-import { DEFAULT_PROMPT_SEARCH_SIZE } from "../config/constants";
-import { errorHandler } from "../middlewares/errorHandler";
-import { CreatePromptImageDto } from "./dtos/prompt-image.dto";
+import { SearchPromptDto } from "../dtos/search-prompt.dto";
+import * as promptService from "../prompt.service";
+import { DEFAULT_PROMPT_SEARCH_SIZE } from "../../config/constants";
+import { errorHandler } from "../../middlewares/errorHandler";
+import { CreatePromptImageDto } from "../dtos/prompt-image.dto";
 
 export const searchPrompts = async (req: Request, res: Response) => {
   try {
