+++ src/prompts/routes/prompt.like.route.ts
@@ -1,6 +1,5 @@
 import { Router } from 'express';
 import { likePrompt, getLikedPrompts } from '../controllers/prompt.like.controller';
-
 const router = Router();
 
 router.post('/prompts/:promptId/likes', likePrompt);
